refactor(ui): split multi-space proposals logic out of Space/Proposals.vue
Space/Proposals.vue is single-space again: no group props, no space
filter, no merged list. Organization/Proposals.vue is now a full view
owning the Spaces filter, merged query, per-row space names, and
selected-space context. Shared wiring moved to useProposalsFilters
(state/labels/space URL sync, one writer) and a ProposalsFilters chips
component used by both views.
